Move near_duplicates detection
