--- freediscovery/dupdet/__init__.py
@@ -1,10 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-
-
-from .simhash import SimhashDuplicates
-from .base import _DuplicateDetectionWrapper
-from .imatch import IMatchDuplicates
