+++ freediscovery/near_duplicates/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+
+from .simhash import SimhashNearDuplicates
+from .imatch import IMatchNearDuplicates
