+++ freediscovery/server/resources.py
@@ -27,7 +27,7 @@ from freediscovery.engine.search import _SearchWrapper
 from ..metrics import (categorization_score,
                        ratio_duplicates_score, f1_same_duplicates_score,
                        mean_duplicates_count_score, _scale_cosine_similarity)
-from ..dupdet import _DuplicateDetectionWrapper
+from freediscovery.engine.near_duplicates import _DuplicateDetectionWrapper
 from freediscovery.engine.email_threading import _EmailThreadingWrapper
 from ..datasets import load_dataset
 from ..exceptions import WrongParameter
