+++ freediscovery/tests/test_integration.py
@@ -11,7 +11,7 @@ import itertools
 
 from freediscovery.text import FeatureVectorizer
 from freediscovery.engine.categorization import _CategorizerWrapper
-from freediscovery.dupdet import _DuplicateDetectionWrapper
+from freediscovery.engine.near_duplicates import _DuplicateDetectionWrapper
 from freediscovery.cluster import _ClusteringWrapper
 from freediscovery.engine.lsi import _LSIWrapper
 from freediscovery.io import parse_ground_truth_file
