+++ freediscovery/tests/test_near_duplicates.py
@@ -1,13 +1,7 @@
 # -*- coding: utf-8 -*-
 
-import os.path
 from unittest import SkipTest
-import numpy as np
 import pytest
-
-from freediscovery.text import FeatureVectorizer
-from .run_suite import check_cache
-
 
 # adapted from https://github.com/seomoz/simhash-py/blob/master/test/test.py
 jabberwocky = '''
@@ -42,19 +36,6 @@ jabberwocky = '''
 jabberwocky_author = ' - Lewis Carroll (Alice in Wonderland)'
 
 
-def fd_setup(**fe_options):
-    basename = os.path.dirname(__file__)
-    cache_dir = check_cache()
-    data_dir = os.path.join(basename, "..", "data", "ds_001", "raw")
-    n_features = 110000
-    fe = FeatureVectorizer(cache_dir=cache_dir)
-    uuid = fe.setup(n_features=n_features, use_hashing=True,
-                    stop_words='english',
-                    **fe_options)
-    fe.ingest(data_dir, file_pattern='.*\d.txt')
-    return cache_dir, uuid, fe.filenames_, fe
-
-
 def test_simhash():
 
     try:
@@ -62,7 +43,7 @@ def test_simhash():
     except ImportError:
         raise SkipTest
     from sklearn.feature_extraction.text import HashingVectorizer
-    from freediscovery.dupdet import SimhashDuplicates
+    from freediscovery.near_duplicates import SimhashNearDuplicates
 
     DISTANCE = 4
 
@@ -73,7 +54,7 @@ def test_simhash():
                           jabberwocky_author,
                           jabberwocky])
 
-    sh = SimhashDuplicates()
+    sh = SimhashNearDuplicates()
     sh.fit(X)
 
     # make sure small changes in the text results in a small number of different bytes
@@ -82,18 +63,18 @@ def test_simhash():
     assert num_differing_bits(*sh._fit_shash[1:3]) >= 20
 
     # same text produces a zero bit difference
-    assert num_differing_bits(*sh._fit_shash[[0,-1]]) == 0
+    assert num_differing_bits(*sh._fit_shash[[0, -1]]) == 0
 
     simhash, cluster_id, dup_pairs = sh.query(distance=DISTANCE, blocks=42)
     assert str(dup_pairs.dtype) == 'uint64'
     assert str(cluster_id.dtype) == 'int64'
     assert str(dup_pairs.dtype) == 'uint64'
 
-    assert simhash[0] == simhash[-1]       # duplicate documents have the same simhash
-    assert cluster_id[0] == cluster_id[-1] # and belong to the same cluster
+    assert simhash[0] == simhash[-1]        # duplicate documents have the same simhash
+    assert cluster_id[0] == cluster_id[-1]  # and belong to the same cluster
 
     for idx, shash in enumerate(simhash):
-        if (shash == simhash).sum() == 1: # ignore duplicates
+        if (shash == simhash).sum() == 1:  # ignore duplicates
             assert sh.get_index_by_hash(shash) == idx
 
     for pairs in dup_pairs:
@@ -104,20 +85,19 @@ def test_simhash():
 def test_imatch(n_rand_lexicons):
 
     from sklearn.feature_extraction.text import TfidfVectorizer
-    from freediscovery.dupdet import IMatchDuplicates
+    from freediscovery.near_duplicates import IMatchNearDuplicates
 
     DISTANCE = 4
 
-    fe = TfidfVectorizer(ngram_range=(4,4), analyzer='word',
+    fe = TfidfVectorizer(ngram_range=(4, 4), analyzer='word',
                          min_df=0.25, max_df=0.75)
 
     X = fe.fit_transform([jabberwocky,
                           jabberwocky + jabberwocky_author,
                           jabberwocky_author,
                           jabberwocky])
-    #print(fe.get_feature_names())
 
-    sh = IMatchDuplicates(n_rand_lexicons=n_rand_lexicons)
+    sh = IMatchNearDuplicates(n_rand_lexicons=n_rand_lexicons)
     sh.fit(X)
 
     assert sh.labels_.shape[0] == X.shape[0]
@@ -131,26 +111,3 @@ def test_imatch(n_rand_lexicons):
     assert sh.labels_[0] == sh.labels_[-1]
 
     # RY: not sure what other tests could be run for I-Match
-
-
-@pytest.mark.parametrize('method, options, fe_options',
-        [['simhash', {'distance': 3}, {} ],
-         ['simhash', {'distance': 10}, {}],
-         ['i-match', {}, {}]])
-def test_dup_detection(method, options, fe_options):
-    if method == 'simhash':
-        try:
-            import simhash
-        except ImportError:
-            raise SkipTest
-    from freediscovery.dupdet import _DuplicateDetectionWrapper
-    cache_dir, uuid, filenames, fe = fd_setup(**fe_options)
-
-    dd = _DuplicateDetectionWrapper(cache_dir=cache_dir, parent_id=uuid)
-    dd.fit(method=method)
-    cluster_id = dd.query(**options)
-    # cannot have more cluster_id than elements in the dataset
-    assert len(np.unique(cluster_id)) <= len(np.unique(filenames))
-
-
-
